+++ wifipumpkin3/core/common/threads.py
@@ -226,7 +226,8 @@ class ProcessHostapd(QObject):
 
 
 class WorkerProcess(QThread):
-    sendOutput = pyqtSignal(str)  
+    sendOutput = pyqtSignal(str)
+
     def __init__(self, command):
         super(WorkerProcess, self).__init__()
         self.process = QProcess(self)
@@ -236,10 +237,12 @@ class WorkerProcess(QThread):
         self.process.terminate()
         if not self.process.waitForFinished(10000):
             self.process.kill()
-    
+
     def start(self):
         self.setupProcess()
 
     def setupProcess(self):
         self.process.setProcessChannelMode(QProcess.MergedChannels)
-        self.process.start(list(self.command.keys())[0], self.command[list(self.command.keys())[0]])
+        self.process.start(
+            list(self.command.keys())[0], self.command[list(self.command.keys())[0]]
+        )